Organize code. Make src/index.ts smaller and more understandable
