+++ src/types/index.d.ts
@@ -1,0 +1,9 @@
+export interface ActivityType {
+  name: string;
+  notes: string;
+  subClassOf: {name: string; href?: string;};
+}
+
+export interface ScrapedVocabulary {
+  activityTypes: ActivityType[];
+}
